python-adk-backend/requirements.txt
google-adk==0.1.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1 